+++ Arenbee/Actors/Players/Default/State/MoveStateMachine.cs
@@ -22,7 +22,7 @@ namespace Arenbee.Actors.Default.State
                 PlayAnimation(AnimationName);
             }
 
-            public override MoveState Update(float delta)
+            public override MoveState Update(double delta)
             {
                 return CheckForTransitions();
             }
@@ -50,7 +50,7 @@ namespace Arenbee.Actors.Default.State
                 Actor.MaxSpeed = Actor.WalkSpeed;
             }
 
-            public override MoveState Update(float delta)
+            public override MoveState Update(double delta)
             {
                 Actor.UpdateDirection();
                 Actor.Move();
@@ -81,7 +81,7 @@ namespace Arenbee.Actors.Default.State
                 Actor.MaxSpeed = Actor.RunSpeed;
             }
 
-            public override MoveState Update(float delta)
+            public override MoveState Update(double delta)
             {
                 Actor.UpdateDirection();
                 Actor.Move();
